refactor: move some types to rlpx.rs
